# Fix gpt optimization