Avoids variable check with unexpected behaviour.

This gives better optimizations for gpt2 on my quadro T1000.
`DEBUG=1 JIT=1 python examples/gpt2.py --model_size gpt2 --timing`
Before:
```
ran model in 39.70 ms, 24.71 ms on GPU, 0.34 GOPS, 0.69 GB, 27.72 GB/s
total 41.13 ms
```
After:
```
ran model in 32.58 ms, 17.46 ms on GPU, 0.34 GOPS, 0.69 GB, 39.23 GB/s
total 34.07 ms
```